Schedule, Proposal dates, Date Dialog, Unit Tests
Added a schedule.txt file to keep track of tasks. Will add future weeks
in next commit.
Updated the dates in the proposal to match actual defense dates.
Changed the date dialog to start off on today's date.
Linked the necessary files properly so the unit testing works on
the emulator.