Simple Web Service?
This code worked the third time I ran it, but threw an exception the
first two times. There is apparently a bug when accessing web services
using Mono for Android. Since this can function, I'll upload it, but I
am still looking for a more sure way of handling it.