MonoDevelop userprefs
As part of wrestling with the seemingly random error mentioned in the
last commit, I installed Xamarin Studio, which created this file when I
opened Zadify with it. The problem magically disappeared later, after I
returned to Visual Studio, but just in case this made a difference
somehow, I'm committing it.